Fix more tests, remove prophecy, remove warnings, rationalise to one PredisMock
