+++ api/app/tests/Unit/PredisMock.php
@@ -6,11 +6,36 @@ namespace Tests\OPG\Digideps\Backend\Unit;
 
 use Predis\Client;
 
-// because Predis\Client uses magic methods which phpunit can't mock
+// create a simple predis Mock to just return keys
 class PredisMock extends Client
 {
-    public function get(): ?string
+    private array $data = [];
+
+    /** @var array<array> $calls */
+    public array $calls = [];
+
+    public function set(string $key, mixed $value): void
     {
-        return '';
+        $this->calls[] = ['set', $key];
+        $this->data[$key] = $value;
+    }
+
+    public function get(string $key): mixed
+    {
+        $this->calls[] = ['get', $key];
+        return $this->data[$key] ?? null;
+    }
+
+    public function expire(string $key, int $seconds): void
+    {
+        $this->calls[] = ['expire', $key, $seconds];
+        if (!isset($this->data[$key])) {
+            throw new \InvalidArgumentException("key $key not set");
+        }
+    }
+
+    public function __call($commandID, $arguments)
+    {
+        throw new \InvalidArgumentException("PredisMock: Method $commandID not implemented");
     }
 }
