+++ api/app/tests/Unit/Service/Audit/AuditEventsTest.php
@@ -5,27 +5,23 @@ declare(strict_types=1);
 namespace Tests\OPG\Digideps\Backend\Unit\Service\Audit;
 
 use OPG\Digideps\Backend\Service\Audit\AuditEvents;
-use PHPUnit\Framework\Attributes\Test;
 use OPG\Digideps\Backend\Service\Time\DateTimeProvider;
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
-use Prophecy\Prophecy\ObjectProphecy;
 
 final class AuditEventsTest extends TestCase
 {
-    use ProphecyTrait;
-
-    /**
-     * @var ObjectProphecy<DateTimeProvider> $dateTimeProvider
-     */
-    private ObjectProphecy $dateTimeProvider;
+    private DateTimeProvider&MockObject $dateTimeProvider;
     private \DateTime $now;
 
     public function setUp(): void
     {
         $this->now = new \DateTime();
-        $this->dateTimeProvider = self::prophesize(DateTimeProvider::class);
-        $this->dateTimeProvider->getDateTime()->shouldBeCalled()->willReturn($this->now);
+        $this->dateTimeProvider = self::createMock(DateTimeProvider::class);
+        $this->dateTimeProvider->expects(self::once())
+            ->method('getDateTime')
+            ->willReturn($this->now);
     }
 
     #[Test]
@@ -41,7 +37,7 @@ final class AuditEventsTest extends TestCase
             'type' => 'audit',
         ];
 
-        $actual = new AuditEvents($this->dateTimeProvider->reveal())->clientArchived(
+        $actual = new AuditEvents($this->dateTimeProvider)->clientArchived(
             'USER_ARCHIVED_CLIENT',
             '12345678',
             new \DateTime('2023-01-01T00:00:00+00:00'),
