+++ api/app/tests/Unit/Service/BruteForce/AttemptsInTimeCheckerTest.php
@@ -7,7 +7,7 @@ namespace Tests\OPG\Digideps\Backend\Unit\Service\BruteForce;
 use OPG\Digideps\Backend\Service\BruteForce\AttemptsInTimeChecker;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
-use Tests\OPG\Digideps\Backend\Unit\Service\PredisMock;
+use Tests\OPG\Digideps\Backend\Unit\PredisMock;
 
 final class AttemptsInTimeCheckerTest extends TestCase
 {
