+++ api/app/tests/Unit/Service/BruteForce/AttemptsIncrementalWaitingCheckerTest.php
@@ -6,7 +6,7 @@ namespace Tests\OPG\Digideps\Backend\Unit\Service\BruteForce;
 
 use OPG\Digideps\Backend\Service\BruteForce\AttemptsIncrementalWaitingChecker;
 use PHPUnit\Framework\TestCase;
-use Tests\OPG\Digideps\Backend\Unit\Service\PredisMock;
+use Tests\OPG\Digideps\Backend\Unit\PredisMock;
 
 final class AttemptsIncrementalWaitingCheckerTest extends TestCase
 {
